aeroguard-backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-pubsub
pandas
requests
